backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase==2.10.0
httpx
pydantic
pydantic-settings
google-genai
playwright
beautifulsoup4
lxml
requests
